Test ws in rust
